+++ generators/view/index.js
@@ -52,6 +52,9 @@
 
 			return this.prompt(prompts).then($response => {
 				this.areYouReady = $response.areYouReady;
+				if (!this.areYouReady) {
+					this.env.error('You are not ready !');
+				}
 			});
 		}